Add files via upload
Given an array of integers, find an equilibrium index in the array. An equilibrium index is an index such that the sum of elements at lower indexes is equal to the sum of elements at higher indexes.
Input:
The first line contains an integer 
𝑛
n representing the number of elements in the array.
The second line contains 
𝑛
n space-separated integers representing the elements of the array.
Output:
If an equilibrium index is found, print the index (1-based index).
If no equilibrium index is found, print "NOT FOUND".
ex: 5
      3 1 5 2 2
output is 3
because 3rd index(1 based index) i.e; 5.Before 3rd index the elements are 3 and 1 and their sum is 4 which is equal to the sum of elements after the 3rd index.